Fix issue with sub-org switch
