--- modules/react-components/src/providers/governance-connector/models/governance-connector.ts
@@ -1,30 +1,0 @@
-/**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-export interface GovernanceCategoryForOrgsInterface {
-	connectors: GovernanceConnectorForOrgsInterface[],
-	id: string,
-	name: string
-}
-
-export interface GovernanceConnectorForOrgsInterface {
-	friendlyName: string,
-	id: string,
-	name: string,
-	properties: string[]
-}
